Move mockExporter to test
